Fixing size warning

+++ tests/search/main.cpp
@@ -57,7 +57,7 @@ int main(int argc, char **argv)
             std::cout << "start " << node.path() << std::endl;
             for (path::Node::iterator iter = node.begin(); iter != node.end(); ++iter)
             {
-                if (size > 0UL && iter->size() < size)
+                if (size > static_cast<size_t>(0) && iter->size() < size)
                     continue;
                 if (countLines)
                     count(std::cout, *iter);
